HACK [REACTOS] Fix warning C4146: unary minus operator applied to unsigned type, result still unsigned
Also make it a critical warning

+++ dll/win32/comctl32/listview.c
@@ -9905,7 +9905,7 @@ static LRESULT LISTVIEW_VScroll(LISTVIEW_INFO *infoPtr, INT nScrollCode,
         break;
 
     case SB_PAGEUP:
-	nScrollDiff = -scrollInfo.nPage;
+	nScrollDiff = -(INT)scrollInfo.nPage;
         break;
 
     case SB_PAGEDOWN:
@@ -10010,7 +10010,7 @@ static LRESULT LISTVIEW_HScroll(LISTVIEW_INFO *infoPtr, INT nScrollCode,
         break;
 
     case SB_PAGELEFT:
-	nScrollDiff = -scrollInfo.nPage;
+	nScrollDiff = -(INT)scrollInfo.nPage;
         break;
 
     case SB_PAGERIGHT:
